+++ spec/docker/run_spec.js
@@ -1,4 +1,4 @@
-import { Q, _, config, defer, async } from 'azk';
+import { Q, _, config, defer, async, utils } from 'azk';
 import docker from 'azk/docker';
 import h from 'spec/spec_helper';
 
@@ -76,17 +76,17 @@ describe("Azk docker module, run method @slow", function() {
     });
   });
 
-  it("should support bind local and remote volumes", function() {
-    var cmd = ["/bin/bash", "-c", "ls -l /azk /app"];
+  it("should support bind volumes", function() {
+    var cmd = ["/bin/bash", "-c", "ls -l /azk"];
     var options = {
       stdout: mocks.stdout, rm: true,
-      volumes: { [__dirname]: "/app" },
-      local_volumes: { '/etc/': "/azk" },
+      volumes: {
+        "/azk": utils.docker.resolvePath(__dirname),
+      },
     }
 
     return docker.run(default_img, cmd, options).then(() => {
       h.expect(outputs.stdout).to.match(/run_spec.js/);
-      h.expect(outputs.stdout).to.match(/hosts/);
     })
   });
 
@@ -128,11 +128,35 @@ describe("Azk docker module, run method @slow", function() {
       var data = yield container.inspect();
       h.expect(data).to.have.deep.property("State.Running", true);
 
-      var log = yield container.logs({stdout: true, stderr: true});
+      var log = "";
+      yield container.logs({stdout: true, stderr: true}).then((stream) => {
+        var stdout = {
+          write(data) { log += data.toString(); }
+        }
+        container.modem.demuxStream(stream, stdout, stdout);
+        return true;
+      });
+
       yield Q.delay(500);
       h.expect(log).to.match(new RegExp(h.escapeRegExp(`AZK_NAME=${data.Name.slice(1)}`), 'm'));
 
       return container.kill();
+    });
+  });
+
+  it("should support extra docker options in start", function() {
+    return async(function* () {
+      var memory = 10 * 1024 * 1024;
+      var cmd    = ["/bin/true"];
+      var opts   = { rm: false, stdout: mocks.stdout, docker: {
+        start : { Privileged: true },
+        create: { Memory: memory },
+      }};
+      var cont = yield docker.run(default_img, cmd, opts);
+      var data = yield cont.inspect();
+
+      h.expect(data).to.have.deep.property('HostConfig.Privileged').and.to.ok;
+      h.expect(data).to.have.deep.property('Config.Memory', memory);
     });
   });
 
@@ -164,6 +188,26 @@ describe("Azk docker module, run method @slow", function() {
       h.expect(data).to.have.deep.property("Annotations.azk.uid").and.match(/^[0-9a-f]+$/);
       h.expect(data).to.have.deep.property("Annotations.azk.key1", "v1");
       h.expect(data).to.have.deep.property("Annotations.azk.key2", "v2");
+    });
+  });
+
+  it("should export annotations to env", function() {
+    return async(function* () {
+      var annotations = { azk: {
+        key1: "v1",
+        key2: "v2",
+        type: "daemon",
+      }};
+      var cmd  = ["/bin/true"];
+      var opts = { rm: false, stdout: mocks.stdout, annotations };
+      var cont = yield docker.run(default_img, cmd, opts);
+      var data = yield cont.inspect();
+
+      var envs = data.Config.Env;
+      h.expect(envs).to.include.something.that.match(/AZK_ENV=test/);
+      h.expect(envs).to.include.something.that.match(/AZK_UID=[0-9a-f]+/);
+      h.expect(envs).to.include.something.that.match(/AZK_KEY1=v1/);
+      h.expect(envs).to.include.something.that.match(/AZK_KEY2=v2/);
     });
   });