+++ spec/system/scale_spec.js
@@ -48,6 +48,36 @@ describe("Azk system class, scale set", function() {
       });
     });
 
+    it("should by default remove the instance after stopping the same", function() {
+      var db = manifest.system('db');
+      return async(this, function* () {
+        var result    = yield db.scale(1);
+        var instances = yield db.instances();
+
+        h.expect(instances).to.length(1);
+        var id = instances[0].Id;
+
+        yield db.scale(0);
+        result = yield docker.findContainer(id);
+        h.expect(result).to.null;
+      });
+    });
+
+    it("should skip remove the instance", function() {
+      var db = manifest.system('db');
+      return async(this, function* () {
+        var result    = yield db.scale(1);
+        var instances = yield db.instances();
+
+        h.expect(instances).to.length(1);
+        var id = instances[0].Id;
+
+        yield db.scale(0, { remove: false });
+        result = yield docker.findContainer(id);
+        h.expect(result).to.not.null;
+      });
+    });
+
     describe("with dependencies is run", function() {
       it("should scale a system with dependencies", function() {
         return async(this, function* () {
@@ -87,6 +117,22 @@ describe("Azk system class, scale set", function() {
 
           icc = yield system.start();
           h.expect(icc).to.equal(3);
+        });
+      });
+
+      it("should not do anything to scale from 0 to 0", function() {
+        return async(this, function* () {
+          var icc, instances, api = manifest.system("api");
+          yield api.stop();
+
+          instances = (yield system.instances())
+          h.expect(instances).to.length(0);
+
+          icc = yield api.scale(0);
+          h.expect(icc).to.equal(0);
+
+          instances = (yield system.instances())
+          h.expect(instances).to.length(0);
         });
       });