+++ src/generator/index.js
@@ -79,8 +79,31 @@ function hash_key(data) {
   return (data || "").match(/^[\w|_]*$/) ? data : `'${data}'`;
 }
 
+function mount(data) {
+  if (_.isString(data)) { return json(data); }
+  var type    = data.type;
+  var options = _.reduce(data, (opts, value, key) => {
+    if (key != 'value' && key != 'type') {
+      opts[key] = value;
+    }
+    return opts;
+  }, {});
+
+  // args
+  var args  = [];
+  if (!_.isUndefined(data.value)) { args.push(data.value) };
+  if (!_.isEmpty(options)) args.push(options);
+  args = _.map(args, (arg) => { return json(arg); });
+
+  switch(data.type) {
+    default:
+      return `${data.type}(${args.join(', ')})`;
+  }
+}
+
 Handlebars.registerHelper('json', json);
 Handlebars.registerHelper('hash_key', hash_key);
+Handlebars.registerHelper('mount', mount);
 
 export { example_system };